Title: 최솟값 3개, Time: 579ms, Memory: 22MB, Status: Passed - CodeTree

+++ 240314/최솟값 3개/top-3-smallest-number.java
@@ -8,21 +8,21 @@ import java.util.StringTokenizer;
 public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        PriorityQueue<Integer> pq = new PriorityQueue<>();
+        PriorityQueue<Long> pq = new PriorityQueue<>();
         int n = Integer.parseInt(br.readLine());
         StringBuilder sb = new StringBuilder();
 
         StringTokenizer st = new StringTokenizer(br.readLine());
 
         for (int i = 0; i < n; i++) {
-            pq.add(Integer.parseInt(st.nextToken()));
+            pq.add(Long.parseLong(st.nextToken()));
 
             if (pq.size() < 3) {
                 sb.append(-1).append("\n");
             } else {
-                int a = pq.poll();
-                int b = pq.poll();
-                int c = pq.poll();
+                long a = pq.poll();
+                long b = pq.poll();
+                long c = pq.poll();
 
                 sb.append(a * b * c).append("\n");
 
@@ -31,8 +31,6 @@ public class Main {
                 pq.add(c);
             }
         }
-
         System.out.println(sb);
-
     }
 }
